requirements.txt
python-dotenv>=1.1.1
unstructured[pdf,local-inference]>=0.18.15
pdf2image>=1.17.0
pytesseract>=0.3.13
tesseract>=0.1.3
poppler-utils>=0.1.0
huggingface_hub[hf_xet]>=0.36.0
